added navigation menu
