--- jtds-1.2.8/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('jtds-1.2.8.jar'))
